zfs: handle forced-export/destroy when suspended.
Add a new txg_how flag of TXG_NOSUSPEND, which indicates the call wants to
be notified (with an error) if its request encounters a pool suspension event.

Add a new mechanism for forced-export/destroy to kill all potentially
long-running activities, in particular send & receive streams.  Use a
'spa killer' barrier in combination with a cancellation mechanism which
closes file descriptors associated with those streams.  Existing mechanisms
handle rejecting new ioctl and VOP requests once a pool is so marked.

On FreeBSD & illumos, fo_close() performs file handle close.  Implement a
SPL shim to cover this.