send/recv: fix cancellation mechanism
This relies on earlier changes, which fix associated bugs.

As part of this change, introduce the notion of cancelling a bqueue, which
otherwise is uninteruptible and could hang if the user is waiting on one end
or the other.  To tolerate this, in the error case, instead of iterating
until the eos marker is seen, just iterate until dequeue returns NULL, which
it can only if cancelled.  Update all callers accordingly.

Leverage thread_signal() to signal the initiating thread of a send/recv;
change their basic functions so they can be interrupted with a signal, and
so they're responsible for pushing their worker thread to completion.
Signal interruption also interrupts any pipe i/o with an error.